refactor: add publication modal
